Move all configuration and properties classes to config package
