+++ taco-cloud/src/main/java/by/kiselevich/tacocloud/config/OrderProps.java
@@ -1,4 +1,4 @@
-package by.kiselevich.tacocloud.props;
+package by.kiselevich.tacocloud.config;
 
 import lombok.Data;
 import org.springframework.boot.context.properties.ConfigurationProperties;
